created a edit button to edit form by frtch data

+++ src/services/apiCabins.js
@@ -1,48 +1,57 @@
 import supabase, { supabaseUrl } from "./supabase";
 
 export async function getCabins() {
-  let { data, error } = await supabase.from("cabins").select("*");
+  const { data, error } = await supabase.from("cabins").select("*");
 
   if (error) {
-    console.log(error);
-    throw new Error("couldn't get data");
+    console.error(error);
+    throw new Error("Cabins could not be loaded");
   }
+
   return data;
 }
 
-export async function createCabin(newCabin) {
-  //Create cabin
+export async function createEditCabin(newCabin, id) {
+  const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
 
-  const imageName = `${Math.random()}
-  -${
-    newCabin.image.name
-  }`.replaceAll("/", "");
-  const imagePath = `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
-  
+  const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
+    "/",
+    ""
+  );
+  const imagePath = hasImagePath
+    ? newCabin.image
+    : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
 
-  //https://wendjyexikqtujaiqage.supabase.co/storage/v1/object/public/cabin-images/cabin-001.jpg
+  // 1. Create/edit cabin
+  let query = supabase.from("cabins");
 
-  const { data, error } = await supabase
-    .from("cabins")
-    .insert([{ ...newCabin, image: imagePath }])
-    .select();
+  // A) CREATE
+  if (!id) query = query.insert([{ ...newCabin, image: imagePath }]);
+
+  // B) EDIT
+  if (id) query = query.update({ ...newCabin, image: imagePath }).eq("id", id);
+
+  const { data, error } = await query.select().single();
 
   if (error) {
-    console.log(error);
+    console.error(error);
     throw new Error("Cabin could not be created");
   }
-  //upload images
+
+  // 2. Upload image
+  if (hasImagePath) return data;
 
   const { error: storageError } = await supabase.storage
     .from("cabin-images")
     .upload(imageName, newCabin.image);
 
-  //Delete cabin if there is an error to upload image
-
+  // 3. Delete the cabin IF there was an error uplaoding image
   if (storageError) {
     await supabase.from("cabins").delete().eq("id", data.id);
-    console.log(storageError);
-    throw new Error("Image could not be uploaded cabin is not created");
+    console.error(storageError);
+    throw new Error(
+      "Cabin image could not be uploaded and the cabin was not created"
+    );
   }
 
   return data;
@@ -52,7 +61,9 @@ export async function deleteCabin(id) {
   const { data, error } = await supabase.from("cabins").delete().eq("id", id);
 
   if (error) {
+    console.error(error);
     throw new Error("Cabin could not be deleted");
   }
+
   return data;
 }
